fixup! fix(ivy): removing unnecessary assert in getOrCreateNodeInjector function
